+++ src/s4u/s4u_Io.cpp
@@ -46,7 +46,7 @@ Io* Io::start()
   return this;
 }
 
-int Io::wait_any_for(const std::vector<IoPtr>& ios, double timeout)
+ssize_t Io::wait_any_for(const std::vector<IoPtr>& ios, double timeout)
 {
   std::vector<kernel::activity::IoImpl*> rios(ios.size());
   std::transform(begin(ios), end(ios), begin(rios),
@@ -54,7 +54,7 @@ int Io::wait_any_for(const std::vector<IoPtr>& ios, double timeout)
 
   kernel::actor::ActorImpl* issuer = kernel::actor::ActorImpl::self();
   kernel::actor::IoWaitanySimcall observer{issuer, rios, timeout};
-  int changed_pos = kernel::actor::simcall_blocking(
+  ssize_t changed_pos = kernel::actor::simcall_blocking(
       [&observer] {
         kernel::activity::IoImpl::wait_any_for(observer.get_issuer(), observer.get_ios(), observer.get_timeout());
       },
