+++ teshsuite/s4u/activity-lifecycle/testing_test-wait.cpp
@@ -62,14 +62,14 @@ template <int Duration, typename Activity> bool tester_wait_any(const Activity& 
   try {
     std::vector<Activity> activities = {activity};
     XBT_DEBUG("calling wait_any_for(%f)", duration);
-    int index = Activity::element_type::wait_any_for(activities, duration);
+    ssize_t index = Activity::element_type::wait_any_for(activities, duration);
     if (index == -1) {
       XBT_DEBUG("wait_any_for() timed out");
       INFO("wait_any_for() timeout should expire at expected date: " << timeout);
       REQUIRE(simgrid::s4u::Engine::get_clock() == Approx(timeout));
       ret = false;
     } else {
-      XBT_DEBUG("wait_any_for() returned index %d", index);
+      XBT_DEBUG("wait_any_for() returned index %zd", index);
       REQUIRE(index == 0);
       ret = true;
     }
